build: ported article store to Pinia
